+++ src/jsc/ipc.rs
@@ -454,7 +454,7 @@ mod advanced {
 mod json {
     use super::*;
 
-    extern "C" fn json_ipc_data_string_free_cb(context: *mut bool, _: *mut c_void, _: u32) {
+    extern "C" fn json_ipc_data_string_free_cb(context: *mut bool, _: *mut c_void, _: usize) {
         // SAFETY: context points to `was_ascii_string_freed` on the caller's stack,
         // kept alive across the deref/defer block in decode_ipc_message.
         unsafe { *context = true };
